Re-introduce hadoop credentials support when specifying credentials (#129)
Re-introduce Hadoop Credentials support when specifying credentials directly via Hadoop Configuration.

Was reverted in commit 2b28ab1f50dbb2d8e4a48ae30496451dd41af676 to
maintain Hadoop 1 compatibility.
