+++ util-hadoop/src/main/hadoop2/java/com/google/cloud/hadoop/util/ConfigurationEntriesAdapterSpecific.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2018 Google Inc. All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software distributed under the
+ * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.cloud.hadoop.util;
+
+import java.io.IOException;
+import org.apache.hadoop.conf.Configuration;
+
+abstract class ConfigurationEntriesAdapterSpecific extends ConfigurationEntriesAdapterGeneric {
+
+  public ConfigurationEntriesAdapterSpecific(Configuration config) {
+    super(config);
+  }
+
+  @Override
+  public String getPassword(String key) {
+    try {
+      char[] val = config.getPassword(key);
+      return val == null ? null : String.valueOf(val);
+    } catch (IOException e) {
+      throw new RuntimeException(e);
+    }
+  }
+}
